Matrix inversion and move precision from f32 to f64

+++ src/play.rs
@@ -29,18 +29,18 @@ impl Projectile {
         Projectile { position, velocity }
     }
 
-    pub fn get_x(&self) -> f32 {
+    pub fn get_x(&self) -> f64 {
         self.position.x
     }
 
-    pub fn get_y(&self) -> f32 {
+    pub fn get_y(&self) -> f64 {
         self.position.y
     }
 
     pub fn to_canvas_coordinate(&self, canvas: &Canvas) -> Tuple {
         Tuple::new_point(
             self.position.x,
-            canvas.height() as f32 - self.position.y,
+            canvas.height() as f64 - self.position.y,
             self.position.z,
         )
     }
